requirements.txt
# AWS SDK and Cloud Storage
boto3==1.34.0
botocore==1.34.0
s3fs==2024.2.0

# Data Processing and Columnar Format
pandas==2.2.0
pyarrow==15.0.0

# Streaming (Kafka/Redpanda)
kafka-python==2.0.2

# Visualization and Dashboarding
streamlit==1.31.0
plotly==5.18.0

# Utilities
python-dateutil==2.8.2